backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
joblib==1.3.1
requests==2.31.0
# Optional for LSTM training: install tensorflow in a compatible Python environment.
psycopg2-binary==2.9.7
